+++ tests/e2e/utils/src/components.js
@@ -368,7 +368,35 @@ const createSimpleOrder = async ( orderStatus = 'Pending payment' ) => {
 
 	// Verify
 	await expect( page ).toMatchElement( '#message', { text: 'Order updated.' } );
+
+	const variablePostId = await page.$( '#post_ID' );
+	let variablePostIdValue = ( await ( await variablePostId.getProperty( 'value' ) ).jsonValue() );
+	return variablePostIdValue;
 };
+
+/**
+ * Adds a product to an order in the StoreOwnerFlow.
+ *
+ * @param orderId ID of the order to add the product to.
+ * @param productName Name of the product being added to the order.
+ */
+const addProductToOrder = async ( orderId, productName ) => {
+	await StoreOwnerFlow.goToOrder( orderId );
+
+	// Add a product to the order
+	await expect( page ).toClick( 'button.add-line-item' );
+	await expect( page ).toClick( 'button.add-order-item' );
+	await page.waitForSelector( '.wc-backbone-modal-header' );
+	await expect( page ).toClick( '.wc-backbone-modal-content .wc-product-search' );
+	await expect( page ).toFill( '#wc-backbone-modal-dialog + .select2-container .select2-search__field', productName );
+	await expect( page ).toClick( 'li[aria-selected="true"]' );
+	await page.click( '.wc-backbone-modal-content #btn-ok' );
+
+	await uiUnblocked();
+
+	// Verify the product we added shows as a line item now
+	await expect( page ).toMatchElement( '.wc-order-item-name', { text: productName } );
+}
 
 export {
 	completeOnboardingWizard,
@@ -376,4 +404,5 @@ export {
 	createVariableProduct,
 	createSimpleOrder,
 	verifyAndPublish,
+	addProductToOrder,
 };